pyproject.toml
[project]
name = "userservice"
version = "0.1.0"
description = ""
authors = [
    {name = "defuzia",email = "zhyvhyn.roman@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "sqlalchemy (>=2.0.39,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.15.1,<2.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "aio-pika (>=9.5.5,<10.0.0)",
    "greenlet (>=3.1.1,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "punq (>=0.7.0,<0.8.0)",
    "pyjwt[crypto] (>=2.10.1,<3.0.0)",
    "orjson (>=3.10.16,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "redis (>=5.2.1,<6.0.0)",
    "aiobotocore (>=2.21.1,<3.0.0)",
    "types-aiobotocore[full] (>=2.21.1,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
